Decouple header files and source files
